+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - Fix incorrect handing of network error response (#944)
 - Show an error when adding a layer with a duplicate ID
+- Replace deprecated `ReactDOM.render` usage with `createRoot`
 - _...Add new stuff here..._
 
 ## 2.1.1